+++ assets/i18n/kr.i18n.yaml
@@ -326,14 +326,15 @@ multi_sig_bsms_screen:
     title: "지갑 상세 정보"
     info_copied: "지갑 상세 정보가 복사됐어요"
   title: "지갑 설정 정보"
-  text1: "안전한 다중 서명 지갑 관리를 위한 표준에 따라 지갑 설정 정보를 관리하고 공유합니다."
-  text2: "모든 키가 볼트에 저장되어 있습니다."
-  text3: "같은 키를 보관하고 있는 다른 볼트에서도 이 QR을 읽어 다중 서명 지갑을 추가할 수 있습니다."
-  text4: "이 다중 서명 지갑에 지정된 **$gen** 키의 니모닉 문구는 현재 다른 볼트에 있습니다."
-  text5: "**$gen** 키 보관 지갑 - **다중 서명 지갑 가져오기**에서 아래 QR 코드를 읽어 주세요. 다중 서명 트랜잭션에 **$gen** 키로 서명하기 위해 이 절차가 반드시 필요합니다."
-  gen1: "$first번"
-  gen2: "$first번과 $last번"
-  gen3: "$first번 또는 $last번"
+  guide:
+    text1: "안전한 다중 서명 지갑 관리를 위한 표준에 따라 지갑 설정 정보를 관리하고 공유합니다."
+    text2: "모든 키가 볼트에 저장되어 있습니다."
+    text3: "같은 키를 보관하고 있는 다른 볼트에서도 이 QR을 읽어 다중 서명 지갑을 추가할 수 있습니다."
+    text4: "이 다중 서명 지갑에 지정된 **$gen** 키의 니모닉 문구는 현재 다른 볼트에 있습니다."
+    text5: "**$gen** 키 보관 지갑 - **다중 서명 지갑 가져오기**에서 아래 QR 코드를 읽어 주세요. 다중 서명 트랜잭션에 **$gen** 키로 서명하기 위해 이 절차가 반드시 필요합니다."
+  first_key: "$first번"
+  first_and_last_keys: "$first번과 $last번"
+  first_or_last_key: "$first번 또는 $last번"
   view_detail: "상세 정보 보기"
 multi_sig_memo_bottom_sheet:
   imported_wallet_memo: "외부 지갑 메모"
