+++ lib/screens/vault_menu/info/mnemonic_view_screen.dart
@@ -22,16 +22,16 @@ class _MnemonicViewScreen extends State<MnemonicViewScreen> {
   bool _isPressed = false;
   bool _isLoading = true;
   List<Widget> _passphraseGridItems = [];
-  late WalletProvider _vaultModel;
+  late WalletProvider _walletProvider;
   String? mnemonic;
   String? passphrase;
 
   @override
   void initState() {
     super.initState();
-    _vaultModel = Provider.of<WalletProvider>(context, listen: false);
+    _walletProvider = Provider.of<WalletProvider>(context, listen: false);
     WidgetsBinding.instance.addPostFrameCallback((_) {
-      _vaultModel.getSecret(widget.walletId).then((secret) async {
+      _walletProvider.getSecret(widget.walletId).then((secret) async {
         await Future.delayed(const Duration(seconds: 1));
         setState(() {
           mnemonic = secret.mnemonic;
